+++ packages/documentation/.storybook/blocks/footer.tsx
@@ -1,26 +1,78 @@
 import React from 'react';
 
-interface Developer {
+interface TeamMember {
   name: string;
   title: string;
-  avatar: string;
+  githubImageId: string;
+  githubUsername: string;
 }
 
-const DEVELOPERS: Developer[] = [
+const TEAM_MEMBERS: TeamMember[] = [
   {
     name: 'Philipp Gfeller',
-    title: 'Lead UI Developer',
-    avatar: 'https://avatars.githubusercontent.com/u/1659006?v=4',
+    title: 'Context switcher',
+    githubImageId: '1659006',
+    githubUsername: 'gfellerph',
   },
   {
     name: 'Alizé Debray',
-    title: 'UI Developer',
-    avatar: 'https://avatars.githubusercontent.com/u/33580481?v=4',
+    title: 'UI Mademoiselle',
+    githubImageId: '33580481',
+    githubUsername: 'alizedebray',
   },
   {
     name: 'Oliver Schürch',
+    title: 'Señor Developer',
+    githubImageId: '9716662',
+    githubUsername: 'oliverschuerch',
+  },
+  {
+    name: 'Myrta Sakellariou',
+    title: 'Mama Pigxel',
+    githubImageId: '66249294',
+    githubUsername: 'myrta2302',
+  },
+  {
+    name: 'Lea Gardavaud',
     title: 'UI Developer',
-    avatar: 'https://avatars.githubusercontent.com/u/9716662?v=4',
+    githubImageId: '183501002',
+    githubUsername: 'leagrdv',
+  },
+  {
+    name: 'Alona Zherdetska',
+    title: 'Component Fairy',
+    githubImageId: '138328641',
+    githubUsername: 'alionazherdetska',
+  },
+  {
+    name: 'Tim Schär',
+    title: 'UI Developer',
+    githubImageId: '59233938',
+    githubUsername: 'schaertim',
+  },
+  {
+    name: 'Rouven Steiger',
+    title: 'Art Guy EDK',
+    githubImageId: '104423005',
+    githubUsername: 'rouvenpost',
+  },
+  {
+    name: 'Christian Zundel',
+    title: 'UX Designer',
+    githubImageId: '119845956',
+    githubUsername: 'Cian77',
+  },
+  {
+    name: 'Alessio Travaglini',
+    title: 'Token Orchestrator',
+    githubImageId: '158268546',
+    githubUsername: 'Vandapanda',
+  },
+  {
+    name: 'Romain Veya',
+    title: 'UI Developer',
+    githubImageId: '111903046',
+    githubUsername: 'veyaromain',
   },
 ];
 
@@ -43,24 +95,29 @@ export default (params: { pathToStoryFile?: String }) => (
       <div className="bg-light">
         <div className="container">
           <div className="pt-big-r pb-big-r">
-            <h2 className="mt-0">Support</h2>
-            <div className="d-flex flex-wrap mt-huge-r mb-huge-r profile-list">
-              {DEVELOPERS.sort(() => (Math.random() > 0.5 ? 1 : -1)).map((developer, index) => (
-                <article key={index} className="avatar">
-                  <img className="profile-picture" src={developer.avatar} alt="" />
-                  <div>
-                    <p>
-                      <strong>{developer.name}</strong>
-                    </p>
-                    <p>{developer.title}</p>
-                  </div>
-                </article>
+            <h2 className="mt-0">Design System Team Members</h2>
+            <ul className="list-profile">
+              {TEAM_MEMBERS.sort(() => (Math.random() > 0.5 ? 1 : -1)).map((developer, index) => (
+                <li key={index}>
+                  <a className="avatar" href={`https://github.com/${developer.githubUsername}`}>
+                    <div className="avatar-image">
+                      <img
+                        src={`https://avatars.githubusercontent.com/u/${developer.githubImageId}?v=4`}
+                        alt={developer.name}
+                      />
+                    </div>
+                    <div className="avatar-description">
+                      <span>{developer.title}</span>
+                      <span>{developer.name}</span>
+                    </div>
+                  </a>
+                </li>
               ))}
-            </div>
+            </ul>
             <div className="row mt-regular-r">
               <div className="col-12 col-rg-auto mt-regular-r">
                 <a
-                  className="btn-primary btn btn-rg btn-animated"
+                  className="btn-primary btn"
                   href="https://github.com/swisspost/design-system/issues"
                   rel="noopener"
                 >