+++ packages/migrations/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @swisspost/design-system-migrations
 
+## 9.0.0-next.3
+
+## 9.0.0-next.2
+
 ## 9.0.0-next.1
 
 ## 9.0.0-next.0